move commands to output contribution
