+++ src/vs/editor/contrib/wordOperations/test/browser/wordOperations.test.ts
@@ -15,7 +15,7 @@ import { Selection } from '../../../../common/core/selection.js';
 import { ILanguageService } from '../../../../common/languages/language.js';
 import { ILanguageConfigurationService } from '../../../../common/languages/languageConfigurationRegistry.js';
 import { ViewModel } from '../../../../common/viewModel/viewModelImpl.js';
-import { CursorWordAccessibilityLeft, CursorWordAccessibilityLeftSelect, CursorWordAccessibilityRight, CursorWordAccessibilityRightSelect, CursorWordEndLeft, CursorWordEndLeftSelect, CursorWordEndRight, CursorWordEndRightSelect, CursorWordLeft, CursorWordLeftSelect, CursorWordRight, CursorWordRightSelect, CursorWordStartLeft, CursorWordStartLeftSelect, CursorWordStartRight, CursorWordStartRightSelect, DeleteInsideWord, DeleteWordEndLeft, DeleteWordEndRight, DeleteWordLeft, DeleteWordRight, DeleteWordStartLeft, DeleteWordStartRight } from '../../browser/wordOperations.js';
+import { CursorWordEndLeft, CursorWordEndLeftSelect, CursorWordEndRight, CursorWordEndRightSelect, CursorWordLeft, CursorWordLeftSelect, CursorWordRight, CursorWordRightSelect, CursorWordStartLeft, CursorWordStartLeftSelect, CursorWordStartRight, CursorWordStartRightSelect, DeleteInsideWord, DeleteWordEndLeft, DeleteWordEndRight, DeleteWordLeft, DeleteWordRight, DeleteWordStartLeft, DeleteWordStartRight } from '../../browser/wordOperations.js';
 import { deserializePipePositions, serializePipePositions, testRepeatedActionAndExtractPositions } from './wordTestUtils.js';
 import { createCodeEditorServices, instantiateTestCodeEditor, withTestCodeEditor } from '../../../../test/browser/testCodeEditor.js';
 import { instantiateTextModel } from '../../../../test/common/testTextModel.js';
@@ -35,10 +35,6 @@ suite('WordOperations', () => {
 	const _cursorWordStartRightSelect = new CursorWordStartRightSelect();
 	const _cursorWordEndRightSelect = new CursorWordEndRightSelect();
 	const _cursorWordRightSelect = new CursorWordRightSelect();
-	const _cursorWordAccessibilityLeft = new CursorWordAccessibilityLeft();
-	const _cursorWordAccessibilityLeftSelect = new CursorWordAccessibilityLeftSelect();
-	const _cursorWordAccessibilityRight = new CursorWordAccessibilityRight();
-	const _cursorWordAccessibilityRightSelect = new CursorWordAccessibilityRightSelect();
 	const _deleteWordLeft = new DeleteWordLeft();
 	const _deleteWordStartLeft = new DeleteWordStartLeft();
 	const _deleteWordEndLeft = new DeleteWordEndLeft();
@@ -72,12 +68,6 @@ suite('WordOperations', () => {
 	}
 	function cursorWordLeft(editor: ICodeEditor, inSelectionMode: boolean = false): void {
 		runEditorCommand(editor, inSelectionMode ? _cursorWordLeftSelect : _cursorWordLeft);
-	}
-	function cursorWordAccessibilityLeft(editor: ICodeEditor, inSelectionMode: boolean = false): void {
-		runEditorCommand(editor, inSelectionMode ? _cursorWordAccessibilityLeft : _cursorWordAccessibilityLeftSelect);
-	}
-	function cursorWordAccessibilityRight(editor: ICodeEditor, inSelectionMode: boolean = false): void {
-		runEditorCommand(editor, inSelectionMode ? _cursorWordAccessibilityRightSelect : _cursorWordAccessibilityRight);
 	}
 	function cursorWordStartLeft(editor: ICodeEditor, inSelectionMode: boolean = false): void {
 		runEditorCommand(editor, inSelectionMode ? _cursorWordStartLeftSelect : _cursorWordStartLeft);
@@ -504,34 +494,6 @@ suite('WordOperations', () => {
 			ed => moveWordStartRight(ed),
 			ed => ed.getPosition()!,
 			ed => ed.getPosition()!.equals(new Position(2, 12))
-		);
-		const actual = serializePipePositions(text, actualStops);
-		assert.deepStrictEqual(actual, EXPECTED);
-	});
-
-	test('cursorWordAccessibilityLeft', () => {
-		const EXPECTED = ['|   /* |Just |some   |more   |text |a+= |3 +|5-|3 + |7 */  '].join('\n');
-		const [text,] = deserializePipePositions(EXPECTED);
-		const actualStops = testRepeatedActionAndExtractPositions(
-			text,
-			new Position(1000, 1000),
-			ed => cursorWordAccessibilityLeft(ed),
-			ed => ed.getPosition()!,
-			ed => ed.getPosition()!.equals(new Position(1, 1))
-		);
-		const actual = serializePipePositions(text, actualStops);
-		assert.deepStrictEqual(actual, EXPECTED);
-	});
-
-	test('cursorWordAccessibilityRight', () => {
-		const EXPECTED = ['   /* |Just |some   |more   |text |a+= |3 +|5-|3 + |7 */  |'].join('\n');
-		const [text,] = deserializePipePositions(EXPECTED);
-		const actualStops = testRepeatedActionAndExtractPositions(
-			text,
-			new Position(1, 1),
-			ed => cursorWordAccessibilityRight(ed),
-			ed => ed.getPosition()!,
-			ed => ed.getPosition()!.equals(new Position(1, 50))
 		);
 		const actual = serializePipePositions(text, actualStops);
 		assert.deepStrictEqual(actual, EXPECTED);
